Rename files containing list and map views
